Initial generic helpers and config block helpers
These helpers are to make config block (not tree or single value) type
resources very easy to add.

When saying config block what is meant is several config keys under the
same path, any sub path will be ignored. This is usefull when you must
configre more than one parameter at the same time to have a valid vyos
commit, or if you do not wish for any sub paths to be part of the
resource.

See "firewall rule set" as an example where we do not wish to
affect the sub paths for all the firewall rules in the set.