Simplify adding resources, some bug fixes, some refactoring
It is now possible to add resources with a small function inline that
proxyies to the full schema based resource functions, see several
resources such as firewall rule.

Fixed some bugs with reading of resources when the ID was known but not
the parameters (used during import). Moved several untested resources to
the simplified way of doing things, fixing multiple bugs.

Improved some output when trying to apply over existing remote resource,
requesting an import, and providing the resource id to use.